[SYCL][FPGA][NFC] Refactor of memory attribute tests
This patch

i.updates tests for memory attributes below:

1. [[intel::numbanks()]]
2. [[intel::force_pow2_depth()]]
3. [[intel::bankwidth()]]
4. [[intel::max_replicates()]]
5. [[intel::bank_bits()]]
6. [[intel::private_copies()]]
7. [[intel::merge("mrg1", "depth")]]
8. [[intel::merge("mrg2", "width")]]
9. [[intel::fpga_register]]
10. [[intel::fpga_memory("MLAB")]]
11. [[intel::fpga_memory("BLOCK_RAM")]]
12. [[intel::doublepump]]
13. [[intel::singlepump]]
14. [[intel::simple_dual_port]]

ii. separates AST and diagnostics to make it easier to read/follow-up.
ii. removes duplicate test cases.
iii. no compiler changes.
